Add Concentration component
