+++ components/Concentrations/config.ts
@@ -1,0 +1,320 @@
+const concentrations: IConcentration[] = [
+  {
+    name: "Web Foundations",
+    courses: [
+      {
+        name: "WEB 1.0",
+        description: "Web Foundations",
+        instructor: "Meredith",
+        link: "/courses/WEB-1.0",
+      },
+      {
+        name: "WEB 1.1",
+        description: "Web Architecture",
+        instructor: "Meredith",
+        link: "/courses/WEB-1.1",
+      },
+    ],
+  },
+  {
+    name: "Front End Web Concentration",
+    courses: [
+      {
+        name: "FEW 1.2",
+        description: "JavaScript Foundations",
+        instructor: "Mitchell",
+        link: "/courses/FEW-1.2",
+      },
+      {
+        name: "FEW 2.1",
+        description: "Writing JavaScript Libraries",
+        instructor: "Mitchell",
+        link: "/courses/FEW-2.1",
+      },
+      {
+        name: "FEW 2.2",
+        description: "Advanced CSS and Web Design",
+        instructor: "Mitchell",
+        link: "/courses/FEW-2.2",
+      },
+      {
+        name: "FEW 2.3",
+        description: "Single Page Web Applications",
+        instructor: "Mitchell",
+        link: "/courses/FEW-2.3",
+      },
+      {
+        name: "FEW 2.4",
+        description: "Native Development with JavaScript",
+        instructor: "Mitchell",
+        link: "/courses/FEW-2.4",
+      },
+      {
+        name: "FEW 2.5",
+        description: "Data Visualization & Web Graphics",
+        instructor: "Mitchell",
+        link: "/courses/FEW-2.5",
+      },
+      {
+        name: "FEW 2.9",
+        description: "Advanced Technical Seminar (GraphQL)",
+        instructor: "Mitchell",
+        link: "/courses/FEW-2.9",
+      },
+    ],
+  },
+  {
+    name: "Back End Web Concentration",
+    courses: [
+      {
+        name: "BEW 1.2",
+        description: "Authentication, Associations, & Advanced Queries",
+        instructor: "Dani",
+        link: "/courses/BEW-1.2",
+      },
+      {
+        name: "BEW 1.3",
+        description: "Server-Side Architectures & Frameworks",
+        instructor: "Dani",
+        link: "/courses/BEW-1.3",
+      },
+      {
+        name: "BEW 2.1",
+        description: "Advanced Web Patterns in Node.js",
+        instructor: "Dani",
+        link: "/courses/BEW-2.1",
+      },
+      {
+        name: "BEW 2.2",
+        description: "DevOps, Deployment & Containers",
+        instructor: "Dani",
+        link: "/courses/BEW-2.2",
+      },
+      {
+        name: "BEW 2.3",
+        description: "Introduction to Web Security",
+        instructor: "Dani",
+        link: "/courses/BEW-2.3",
+      },
+      {
+        name: "BEW 2.4",
+        description: "Decentralized Apps & Distributed Protocols",
+        instructor: "Dani",
+        link: "/courses/BEW-2.4",
+      },
+      {
+        name: "BEW 2.5",
+        description: "Go - Patterns & Practices in Strongly Typed Ecosystems",
+        instructor: "Dani",
+        link: "/courses/BEW-2.5",
+      },
+    ],
+  },
+  {
+    name: "Mobile Development",
+    courses: [
+      {
+        name: "MOB 1.1",
+        description: "Introduction to Swift",
+        instructor: "Adriana",
+        link: "/courses/MOB-1.1",
+      },
+      {
+        name: "MOB 1.2",
+        description: "Introduction to iOS Development in Swift",
+        instructor: "Adriana",
+        link: "/courses/MOB-1.2",
+      },
+      {
+        name: "MOB 1.3",
+        description: "Dynamic iOS Applications",
+        instructor: "Adriana",
+        link: "/courses/MOB-1.3",
+      },
+      {
+        name: "MOB 2.1",
+        description: "Local Persistence in iOS",
+        instructor: "Adriana",
+        link: "/courses/MOB-2.1",
+      },
+      {
+        name: "MOB 2.2",
+        description: "2D Mobile Game Development",
+        instructor: "Adriana",
+        link: "/courses/MOB-2.2",
+      },
+      {
+        name: "MOB 2.3",
+        description: "Concurrency & Parallelism in iOS",
+        instructor: "Adriana",
+        link: "/courses/MOB-2.3",
+      },
+      {
+        name: "MOB 2.4",
+        description: "Advanced Architectural Patterns in iOS",
+        instructor: "Adriana",
+        link: "/courses/MOB-2.4",
+      },
+      {
+        name: "MOB 2.9",
+        description: "Advanced Technical Seminar (MOB)",
+        instructor: "Adriana",
+        link: "/courses/MOB-2.9",
+      },
+    ],
+  },
+  {
+    name: "Data Science",
+    courses: [
+      {
+        name: "DS 1.1",
+        description: "Data Analysis & Visualization",
+        instructor: "Joseph",
+        link: "/courses/DS-1.1",
+      },
+      {
+        name: "DS 2.1",
+        description: "Machine Learning",
+        instructor: "Joseph",
+        link: "/courses/DS-2.1",
+      },
+      {
+        name: "DS 2.2",
+        description: "Deep Learning",
+        instructor: "Joseph",
+        link: "/courses/DS-2.2",
+      },
+      {
+        name: "DS 2.3",
+        description: "Data Science in Production",
+        instructor: "Joseph",
+        link: "/courses/DS-2.3",
+      },
+      {
+        name: "DS 2.4",
+        description: "Core Applications of Artificial Intelligence",
+        instructor: "Joseph",
+        link: "/courses/DS-2.4",
+      },
+      {
+        name: "DS 2.9",
+        description: "Advanced Technical Seminar (DS)",
+        instructor: "Joseph",
+        link: "/courses/DS-2.9",
+      },
+    ],
+  },
+  {
+    name: "Computer Science (Core)",
+    courses: [
+      {
+        name: "CS 1.0",
+        description: "Introduction to Programming",
+        instructor: "Joi",
+        link: "/courses/CS-1.0",
+      },
+      {
+        name: "CS 1.1",
+        description: "Object Oriented Programming",
+        instructor: "Joi",
+        link: "/courses/CS-1.1",
+      },
+      {
+        name: "CS 1.2",
+        description: "Introduction to Data Structures",
+        instructor: "Joi",
+        link: "/courses/CS-1.2",
+      },
+      {
+        name: "CS 1.3",
+        description: "Core Data Structures and Algorithms",
+        instructor: "Joi",
+        link: "/courses/CS-1.3",
+      },
+      {
+        name: "CS 2.1",
+        description: "Advanced Trees & Sorting Algorithms",
+        instructor: "Joi",
+        link: "/courses/CS-2.1",
+      },
+      {
+        name: "CS 2.2",
+        description: "Advanced Recursion & Graphs",
+        instructor: "Joi",
+        link: "/courses/CS-2.2",
+      },
+    ],
+  },
+  {
+    name: "Software Product Development",
+    courses: [
+      {
+        name: "SPD 1.1",
+        description: "Introduction to Software Development",
+        instructor: "Braus",
+        link: "/courses/SPD-1.1",
+      },
+      {
+        name: "SPD 1.2",
+        description: "Agile Product Development & UI/UX",
+        instructor: "Braus",
+        link: "/courses/SPD-1.2",
+      },
+      {
+        name: "SPD 1.3",
+        description: "Team Software Project",
+        instructor: "Braus",
+        link: "/courses/SPD-1.3",
+      },
+      {
+        name: "SPD 1.41",
+        description: "Engineering Careers 1: Communication & Interviewing",
+        instructor: "Braus",
+        link: "/courses/SPD-1.41",
+      },
+      {
+        name: "SPD 1.5",
+        description: "Engineering Careers 2: Management and Leadership",
+        instructor: "Braus",
+        link: "/courses/SPD-1.5",
+      },
+      {
+        name: "SPD 2.1",
+        description: "Industry Collaboration Project Part 1: Client Work",
+        instructor: "Braus",
+        link: "/courses/SPD-2.1",
+      },
+      {
+        name: "SPD 2.2",
+        description: "Industry Collaboration Project Part 2: Analytics & Growth",
+        instructor: "Braus",
+        link: "/courses/SPD-2.2",
+      },
+      {
+        name: "SPD 2.31",
+        description: "Testing and Architecture",
+        instructor: "Braus",
+        link: "/courses/SPD-2.31",
+      },
+      {
+        name: "SPD 2.41",
+        description: "Onboarding into Companies",
+        instructor: "Braus",
+        link: "/courses/SPD-2.41",
+      },
+    ],
+  },
+  {
+    name: "Miscellaneous",
+    courses: [
+      {
+        name: "QL 1.1",
+        description: "Quantitative Reasoning",
+        instructor: "Kami",
+        link: "/courses/QL-1.1",
+      },
+    ],
+  },
+]
+
+export default concentrations
