+++ types/makeschool.d.ts
@@ -1,0 +1,11 @@
+interface ICourseWebsite {
+  name: string
+  description: string
+  instructor: string
+  link: string
+}
+
+interface IConcentration {
+  name: string
+  courses: ICourseWebsite[]
+}
